Updated Computer Play and end of game detection
Wrote a simple interface to enable testing as I am having a hard time getting the ncurses display working properly.
Reworked the win and cat game detection for more accuracy.
The play algorithm still needs some work as it will make some stupid choices, such as playing in a place which is already blocked.
Worked a bit more in the display but it still needs a bit of work.